+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v1.0.2
+- Rewrite in typescript
 - Trigger resize event on resize without comparing width
 
 ### v1.0.1