+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.0.4
+- Fix resize event trigger only when width is changed.
+
 ### v1.0.2
 - Rewrite in typescript
 - Trigger resize event on resize without comparing width